feat: rename modal states to Pages
